refactor: extract shared types into @zypher/types workspace package
Create a new @zypher/types package as the single source of truth for
type definitions shared across agent, http, and ui packages. This
eliminates the need for dnt/npm publishing of the ui package and removes
type duplication.

- Add packages/types with all shared type definitions
- Refactor agent, http, and ui packages to import from @zypher/types
- Remove @deno/dnt dependency from ui package
