+++ packages/types/deno.json
@@ -1,0 +1,7 @@
+{
+  "name": "@zypher/types",
+  "version": "0.1.0",
+  "exports": {
+    ".": "./mod.ts"
+  }
+}
