+++ packages/ui/use_agent.ts
@@ -2,8 +2,8 @@ import { hexoid } from "hexoid";
 import type { Dispatch, SetStateAction } from "react";
 import { useCallback, useEffect, useRef, useState, useTransition } from "react";
 import type { TaskApiClient, TaskConnection } from "./task_api_client.ts";
-import type { ContentBlock } from "@zypher/agent";
-import type { TaskWebSocketServerMessage } from "@zypher/http";
+import type { ContentBlock } from "@zypher/types";
+import type { TaskWebSocketServerMessage } from "@zypher/types";
 import type { Observable } from "rxjs";
 import { eachValueFrom } from "rxjs-for-await";
 import useSWR, { type KeyedMutator } from "swr";
